util fix return

+++ src/main/java/com/postales/util/Util.java
@@ -4,7 +4,13 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public class Util {
 
-    public static double generarEnteroAleatorio(int minimo, int maximo) {
+    /**
+     * Genera un número entero aleatorio
+     * @param minimo Mínimo valor entero
+     * @param maximo Máximo valor entero
+     * @return Número entero aleatorio
+     */
+    public static int generarEnteroAleatorio(int minimo, int maximo) {
         return ThreadLocalRandom.current().nextInt(minimo, maximo + 1);
     }
